backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
Pillow
numpy
tensorflow
# Add other libraries your model needs, like scikit-learn, etc.